Forced update
The build will not continue because the command `cf install-plugin autopilot` is waiting for confirmation. This commit directly into master, while not to my liking, is a quick fix to get the builds in a good state. 

The command was changed from `cf install-plugin autopilot` to `cf install-plugin autopilot -f`. The `-f` flag will force the installation without waiting for confirmation.