Logging refactor (#255)
**Summary**

Improvement to application logging.

Explain the **motivation** for making this change. What existing problem does the pull request solve?

The logging for the application was a bit hard to follow when queries needed to be done to the logs.

There was a ton of unnecessary inheritance that was eliminated. Serializers were also added to help improve request and response logging

**Test plan (required)**

Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots / videos if the pull request changes UI.